Nebula query engine
The Nebula project (https://github.com/jongpie/NebulaFramework) has added a lot of new features for dynamic querying, but it's currently dependent on the rest of the Nebula framework. The query engine from Nebula has been ported so that it's a freestanding project that can be deployed without the rest of the framework.